requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
dj-database-url==2.2.0
Django==4.2.13
django-appconf==1.0.6
django-compressor==4.4
django-countries==7.6.1
django-environ==0.11.2
django-phonenumber-field==7.3.0
filelock==3.15.1
identify==2.5.36
nodeenv==1.9.1
phonenumbers==8.13.38
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
PyYAML==6.0.1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.5.0
typing_extensions==4.12.2
virtualenv==20.26.2
